Move old regular polygon test to primitive version
